requirements.txt
pillow
pytesseract
appium-python-client